feat: move remaining functions in principal class to the post actions class
