+++ admin/helpers/class-language-helper.php
@@ -31,9 +31,11 @@ class Language_Helper {
   /**
    * Initializes the class with the default language object.
    *
+   * @param string $plugin   The plugin name.
+   *
    * @since 3.0.0
    */
-  public function __construct() {
+  public function __construct( $plugin ) {
     $this->default_lang = (object) array(
       'language_code'  => 'en',
       'locale'         => 'en-us',
@@ -41,6 +43,7 @@ class Language_Helper {
       'display_name'   => 'English',
       'native_name'    => 'English',
     );
+    $this->plugin       = $plugin;
   }
 
   /**
@@ -93,21 +96,21 @@ class Language_Helper {
    * @since 2.0.0
    */
   public function load_languages() {
+    $post_actions = new \ES_Feeder\Post_Actions( $this->plugin );
+
     if ( defined( 'DOING_AJAX' ) && DOING_AJAX ) {
       $this->languages = get_option( 'cdp_languages' );
       if ( $this->languages ) {
         return;
       }
     }
-    global $feeder;
-    if ( ! $feeder ) {
-      return;
-    }
+
     $args = array(
       'method' => 'GET',
       'url'    => 'language',
     );
-    $data = $feeder->es_request( $args );
+
+    $data = $post_actions->request( $args );
     if ( $data && count( $data ) && ! is_string( $data )
         && ( ! is_array( $data ) || ( is_array( $data ) && ( ! array_key_exists( 'error', $data ) || ! $data['error'] ) )
         && ( ! is_object( $data ) || ( is_object( $data ) && ! $data->error ) ) ) ) {
@@ -116,6 +119,7 @@ class Language_Helper {
         $this->languages[ $lang->locale ] = $lang;
       }
     }
+
     update_option( 'cdp_languages', $this->languages );
   }
 
@@ -142,35 +146,45 @@ class Language_Helper {
    */
   public function get_translations( $post_id ) {
     global $wpdb;
+
     if ( ! function_exists( 'icl_object_id' ) ) {
       return array();
     }
+
     $query = "SELECT trid, element_type FROM {$wpdb->prefix}icl_translations WHERE element_id = $post_id";
     $vars  = $wpdb->get_row( $query );
+
     if ( ! $vars || ! $vars->trid || ! $vars->element_type ) {
       return array();
     }
+
     $query        = "SELECT element_id, language_code FROM {$wpdb->prefix}icl_translations WHERE trid = $vars->trid AND element_type = '$vars->element_type' AND element_id != $post_id";
     $results      = $wpdb->get_results( $query );
     $translations = array();
+
     foreach ( $results as $result ) {
       $lang = $this->get_language_by_code( $result->language_code );
       if ( ! $lang ) {
         continue;
       }
       $status = get_post_status( $result->element_id );
-      if ( $status !== 'publish' ) {
+
+      if ( 'publish' !== $status ) {
         continue;
       }
+
       $sync = get_post_meta( $result->element_id, '_iip_index_post_to_cdp_option', true );
+
       if ( 'no' === $sync ) {
         continue;
       }
+
       $translations[] = array(
         'post_id'  => $result->element_id,
         'language' => $lang,
       );
     }
+
     return $translations;
   }
 }
