+++ admin/helpers/class-owner-helper.php
@@ -16,6 +16,17 @@ namespace ES_Feeder\Admin\Helpers;
  */
 class Owner_Helper {
 
+  /**
+   * Initializes the class with the plugin name and version.
+   *
+   * @param string $plugin   The plugin name.
+   *
+   * @since 3.0.0
+   */
+  public function __construct( $plugin ) {
+    $this->plugin = $plugin;
+  }
+
   public $owners;
 
   /**
@@ -26,23 +37,23 @@ class Owner_Helper {
    *
    * @since 2.5.0
    */
-  public static function get_owners() {
+  public function get_owners() {
+    $post_actions = new \ES_Feeder\Post_Actions( $this->plugin );
+
     if ( defined( 'DOING_AJAX' ) && DOING_AJAX ) {
       $owners = get_option( 'cdp_owners' );
       if ( $owners ) {
         return $owners;
       }
     }
-    global $feeder;
-    if ( ! $feeder ) {
-      return array();
-    }
+
     $owners = array( '' );
     $args   = array(
       'method' => 'GET',
       'url'    => 'owner',
     );
-    $data   = $feeder->es_request( $args );
+    $data   = $post_actions->request( $args );
+
     if ( $data && count( $data ) && ! is_string( $data )
       && ( ! is_array( $data ) || ! array_key_exists( 'error', $data ) || ! $data['error'] )
       && ( ! is_object( $data ) || ! $data->error ) ) {
@@ -50,7 +61,9 @@ class Owner_Helper {
         $owners[ $owner->name ] = $owner->name;
       }
     }
+
     update_option( 'cdp_owners', $owners );
+
     return $owners;
   }
