+++ admin/partials/wp-es-feeder-api-view-display.php
@@ -1,16 +1,26 @@
 <?php
-global $feeder, $post;
+
+global $post;
+
+$post_helper = new ES_Feeder\Admin\Helpers\Post_Helper( $this->plugin );
 
 $options = get_option( $this->plugin );
 $es_url  = $options['es_url'] ? $options['es_url'] : null;
 $token   = $options['es_token'];
 
 if ( $post && $post->ID && $es_url && $token ) :
-  $uuid     = $feeder->get_uuid( $post );
-  $endpoint = $es_url . $feeder->get_post_type_label( $post->post_type ) . '/' . $uuid;
+  $uuid     = $post_helper->get_uuid( $post );
+  $endpoint = $es_url . $post_helper->get_post_type_label( $post->post_type ) . '/' . $uuid;
+
   ?>
   <div style="text-align: right;">
-    <a class="button-secondary" href="javascript:void(0);" id="populate_data">Retrieve Data</a>
+    <a
+      class="button-secondary"
+      href="javascript:void(0);"
+      id="populate_data"
+    >
+      <?php esc_html_e( 'Retrieve Data', 'gpalab-feeder' ); ?>
+    </a>
   </div>
   <pre id="es_response"></pre>
   <script type="text/javascript">
@@ -22,8 +32,8 @@ if ( $post && $post->ID && $es_url && $token ) :
         $.ajax({
           type: 'GET',
           dataType: 'JSON',
-          url: '<?php echo $endpoint; ?>',
-          headers: { Authorization: 'Bearer <?php echo $token; ?>'},
+          url: '<?php echo esc_html( $endpoint ); ?>',
+          headers: { Authorization: 'Bearer <?php echo esc_html( $token ); ?>'},
           success: function (result) {
             if (result && result.content) result.content = "OMITTED";
             $('#es_response').html(JSON.stringify(result, null, 2));
